SequenceColor extension - refactor
