+++ src/mol-plugin-ui/sequence/chain.ts
@@ -47,7 +47,7 @@ export class ChainSequenceWrapper extends SequenceWrapper<StructureUnit> {
         return Interval.Empty;
     }
 
-    getLoci(seqIdx: number) {
+    _getLoci(seqIdx: number) {
         return this.loci;
     }
